Added `nor()`, improved `or()`,
Also created a new function `escape_all` to fix the issue of not being able to escape a collection of Into<String> options for the logical functions. This problem caused a panic in the stop words cookbook section because if the result of `or()` was wrapped in square brackets, the hyphens in the text pulled from the stop words library would be interpreted as ranges.